Restructure test directories
Organize tests by type:
- test/unit/index.js - Unit tests (mock-based, no hardware)
- test/integration/index.js - Integration tests (requires hardware)
- test/helpers/mock.js - Mock implementations
- test/types/index.test-d.ts - TypeScript declaration tests

Update package.json scripts to use new paths.

New structure:
```
test/
├── helpers/
│   └── mock.js
├── integration/
│   └── index.js
├── types/
│   └── index.test-d.ts
└── unit/
    └── index.js
```

Closes #59
