+++ package.json
@@ -19,10 +19,10 @@
         "install": "node-gyp rebuild",
         "build": "node-gyp rebuild",
         "rebuild": "node-gyp rebuild",
-        "test": "node test/test.js && node --test test/unit.js",
-        "test:unit": "node --test test/unit.js",
-        "test:integration": "node test/test.js",
-        "test:coverage": "c8 --include=lib/**/*.js node --test test/unit.js",
+        "test": "node test/integration/index.js && node --test test/unit/index.js",
+        "test:unit": "node --test test/unit/index.js",
+        "test:integration": "node test/integration/index.js",
+        "test:coverage": "c8 --include=lib/**/*.js node --test test/unit/index.js",
         "test:types": "tsd",
         "coverage:report": "c8 report --reporter=text --reporter=html"
     },
